+++ src/browsers/firefox.playwright.ts
@@ -42,13 +42,15 @@ export class FirefoxPlaywright extends EventEmitter {
     this.removeAllListeners();
   }
 
-  public keepAlive() {
-    return false;
+  public keepUntil() {
+    return 0;
   }
 
-  public isRunning = (): boolean => this.running;
+  public isRunning(): boolean {
+    return this.running;
+  }
 
-  public close = async (): Promise<void> => {
+  public async close(): Promise<void> {
     if (this.browser) {
       this.logger.trace(
         `Closing ${this.constructor.name} process and all listeners`,
@@ -60,33 +62,34 @@ export class FirefoxPlaywright extends EventEmitter {
       this.browser = null;
       this.browserWSEndpoint = null;
     }
-  };
+  }
 
-  public pages = async (): Promise<[]> => [];
+  public async pages(): Promise<[]> {
+    return [];
+  }
 
-  public getPageId = (): string => {
+  public getPageId(): string {
     throw new ServerError(
       `#getPageId is not yet supported with ${this.constructor.name}.`,
     );
-  };
+  }
 
-  public makeLiveURL = (): void => {
+  public makeLiveURL(): void {
     throw new ServerError(
       `Live URLs are not yet supported with ${this.constructor.name}.`,
     );
-  };
+  }
 
-  public newPage = async (): Promise<Page> => {
+  public async newPage(): Promise<Page> {
     throw new ServerError(
       `Can't create new page with ${this.constructor.name}`,
     );
-  };
+  }
 
-  public launch = async (
+  public async launch(
     options: BrowserServerOptions = {},
     version?: string,
-  ): Promise<playwright.BrowserServer> => {
-
+  ): Promise<playwright.BrowserServer> {
     this.logger.info(`Launching ${this.constructor.name} Handler`);
     const opts = {
       ...options,
@@ -109,11 +112,13 @@ export class FirefoxPlaywright extends EventEmitter {
     this.running = true;
 
     return this.browser;
-  };
+  }
 
-  public wsEndpoint = (): string | null => this.browserWSEndpoint;
+  public wsEndpoint(): string | null {
+    return this.browserWSEndpoint;
+  }
 
-  public publicWSEndpoint = (token: string | null): string | null => {
+  public publicWSEndpoint(token: string | null): string | null {
     if (!this.browserWSEndpoint) {
       return null;
     }
@@ -128,18 +133,18 @@ export class FirefoxPlaywright extends EventEmitter {
     }
 
     return externalURL.href;
-  };
+  }
 
-  public proxyPageWebSocket = async () => {
+  public async proxyPageWebSocket() {
     this.logger.warn(`Not yet implemented in ${this.constructor.name}`);
-  };
+  }
 
-  public proxyWebSocket = async (
+  public async proxyWebSocket(
     req: Request,
     socket: Duplex,
     head: Buffer,
-  ): Promise<void> =>
-    new Promise((resolve, reject) => {
+  ): Promise<void> {
+    return new Promise((resolve, reject) => {
       if (!this.browserWSEndpoint) {
         throw new ServerError(
           `No browserWSEndpoint found, did you launch first?`,
@@ -173,4 +178,5 @@ export class FirefoxPlaywright extends EventEmitter {
         },
       );
     });
+  }
 }